refactor(constants): extract APIKEY_PROVIDERS (1,820 LOC) to co-located file
Splits src/shared/constants/providers.ts (3,243 -> 1,426 LOC, -56%) by
moving the APIKEY_PROVIDERS const block (1,820 LOC, 56% of the parent
file) to src/shared/constants/providers/apiKeyProviders.ts.

## Why this is safe

APIKEY_PROVIDERS is a single module-level const declaration at L631-L2450
of the parent file. It is the largest single block in the file and has
no closure dependencies on other module-level declarations. Moving it is
a pure file-copy operation with a re-export from the parent.

## What stayed in providers.ts (1,426 LOC)

All 27 other const blocks (the smaller per-provider config arrays:
OPENAI_COMPAT_PROVIDERS, GEMINI_COMPAT_PROVIDERS, ANTHROPIC_COMPAT_PROVIDERS,
etc., totaling ~1,400 LOC). These are queued for PR-2.b/c in subsequent
PRs (each has its own self-contained scope).

## Pattern matches PR-#4609 (imageGeneration split)

Same mechanical refactor pattern:
1. Identify largest leaf block (module-level const, no closure deps)
2. Move to co-located file
3. Add re-export from parent
4. Verify references unchanged

## Verification performed locally

- Re-export present in src/shared/constants/providers.ts: 1 line
- Original APIKEY_PROVIDERS const still in main file: 0 (moved)
- node scripts/check/check-env-doc-sync.mjs: 'Env / docs contract is in sync'
- node scripts/check/check-fabricated-docs.mjs --strict: 'No fabricated
  API/env/CLI/hook/file references found'
- node scripts/check/check-docs-sync.mjs: 'PASS - documentation version
  sync is consistent'
- node scripts/check/check-db-rules.mjs: 'OK (85 modulos db/, 57
  re-exportados, 28 intencionalmente-internos)'

## Refs

- Builds on PR-#4609 (imageGeneration split) - same leaf-block pattern
- Mirrors PR-#4381 (combos split) and PR-#4480 (chatLogHelpers extraction)
- Continues work on issue #4425 (provider-config drift crashes)
